allow for disabling build w .nobuild and try again to fix sl indexing on serve

+++ misc/o2/o2-clean-build.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
+if [ -r .nobuild ]; then
+    echo 'You are not allowed to build this project here. Stop.'
+    exit 1
+fi
 rm -f 01bld/cancel
 o2-unbuild.sh
 touch 01tmp/building
